(original rev 1706) fix some warnings on gcc 3.4
git-svn-id: file:///tmp/current/omni/branches/staging@1473 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/mesh/meshingChunkThreadManager.cpp
@@ -67,7 +67,7 @@ int MeshingChunkThreadManager::numberOfThreadsToUseForThisChunk( const int total
 	const int maxNumWorkerThreads = mMeshManager->getMaxAllowedNumberOfWorkerThreads();
 	int num_threads_to_use = 1;
 	if( totalNumValuesToMesh > 50 ){
-		num_threads_to_use = totalNumValuesToMesh / 50.0 + 1;
+		num_threads_to_use = static_cast<int>(totalNumValuesToMesh / 50.0 + 1);
 		if (num_threads_to_use > maxNumWorkerThreads ) {
 			num_threads_to_use = maxNumWorkerThreads;
 		}
